# services/underwriting/openapi.yaml
openapi: 3.0.3
info:
  title: Underwriting Service API
  version: 0.1.0
paths:
  /api/applications/{id}/underwrite:
    post:
      summary: Evaluate underwriting rules
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                monthlyIncome: { type: number }
                existingEmi: { type: number }
                proposedAmount: { type: number }
                tenureMonths: { type: integer }
                annualRate: { type: number }
                propertyValue: { type: number }
                applicantAgeYears: { type: number }
                product:
                  type: object
                  properties:
                    maxFOIR: { type: number }
                    maxLTV: { type: number }
                    maxAgeAtMaturity: { type: integer }
      responses:
        '200':
          description: Decision
          content:
            application/json:
              schema:
                type: object
                properties:
                  decisionId: { type: string, format: uuid }
                  decision: { type: string, enum: [AUTO_APPROVE, REFER, DECLINE] }
                  reasons: { type: array, items: { type: string } }
                  metrics:
                    type: object
                    properties:
                      foir: { type: number }
                      ltv: { type: number }
                      ageAtMaturity: { type: number }
                      proposedEmi: { type: number }
  /api/applications/{id}/decision:
    get:
      summary: Get latest underwriting decision
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Latest decision
  /api/applications/{id}/override/request:
    post:
      summary: Maker creates override request
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                originalDecision: { type: string, enum: [REFER, DECLINE] }
                requestedDecision: { type: string, enum: [AUTO_APPROVE, REFER] }
                justification: { type: string, minLength: 10, maxLength: 1000 }
              required: [originalDecision, requestedDecision, justification]
      responses:
        '201':
          description: Override request created
  /api/applications/{id}/override/{overrideRequestId}/approve:
    post:
      summary: Checker approves override request
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: overrideRequestId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                remarks: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Override approved
  /api/applications/{id}/override/{overrideRequestId}/reject:
    post:
      summary: Checker rejects override request
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: overrideRequestId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                remarks: { type: string, minLength: 10, maxLength: 500 }
              required: [remarks]
      responses:
        '200':
          description: Override rejected
  /api/applications/{id}/override:
    get:
      summary: List override requests for application
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: List of override requests
  /api/override-requests/pending:
    get:
      summary: List pending override requests (for checkers)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated list of pending override requests